+++ src/DbTests/SQL/DynamicSqlSensitiveToParamValues.sql
@@ -1,0 +1,2 @@
+﻿DECLARE @sql NVARCHAR(MAX) = 'SELECT * FROM dbo.Table1 ORDER BY ' + @orderBy
+EXEC sp_executesql @sql, N''